+++ src/components/TextField/TextField.types.ts
@@ -10,6 +10,8 @@ type BaseProps = {
 	name: string;
 	size?: Size;
 	disabled?: boolean;
+	focused?: boolean;
+	multiline?: boolean;
 	placeholder?: string;
 	icon?: SlotProps["icon"];
 	endIcon?: SlotProps["icon"];